Renaming build/eventhandler -> build/buildeventhandler to be more specific, and to be more difficult to type, apparently.
