+++ cmd/hookhandler/main.go
@@ -7,8 +7,8 @@ import (
 	"strings"
 
 	"github.com/gorilla/mux"
-	"github.com/level11consulting/ocelot/build/eventhandler/push/buildjob"
-	"github.com/level11consulting/ocelot/build/eventhandler/push/webhook"
+	"github.com/level11consulting/ocelot/build/buildeventhandler/push/buildjob"
+	"github.com/level11consulting/ocelot/build/buildeventhandler/push/webhook"
 	"github.com/level11consulting/ocelot/client/buildconfigvalidator"
 	"github.com/level11consulting/ocelot/client/newbuildjob"
 	"github.com/level11consulting/ocelot/models/pb"
