+++ router/hookhandler/hookhandler.go
@@ -5,8 +5,8 @@ import (
 	"net/http"
 	"strings"
 
-	"github.com/level11consulting/ocelot/build/eventhandler/push/buildjob"
-	"github.com/level11consulting/ocelot/build/eventhandler/push/webhook"
+	"github.com/level11consulting/ocelot/build/buildeventhandler/push/buildjob"
+	"github.com/level11consulting/ocelot/build/buildeventhandler/push/webhook"
 	"github.com/level11consulting/ocelot/build/vcshandler"
 	"github.com/level11consulting/ocelot/client/newbuildjob"
 	"github.com/level11consulting/ocelot/client/werkerinterface"
